fix: Update import paths for bind_dims and unbind_dims in test files
