+++ effectful/handlers/jax/__init__.py
@@ -8,6 +8,8 @@ except ImportError:
 # side effect: register defdata for jax.Array
 import effectful.handlers.jax.terms  # noqa: F401
 
+from .handlers import bind_dims as bind_dims
 from .handlers import jax_getitem as jax_getitem
 from .handlers import jit as jit
 from .handlers import sizesof as sizesof
+from .handlers import unbind_dims as unbind_dims
